ai_agent_project/requirements.txt
Django>=4.2
djangorestframework
requests
python-dotenv
faiss-cpu
sentence-transformers
numpy
gunicorn